+++ api/service/listen_task.go
@@ -3,13 +3,18 @@ package service
 import (
 	"encoding/json"
 
+	"github.com/mesg-foundation/core/database/services"
 	"github.com/mesg-foundation/core/execution"
 	"github.com/mesg-foundation/core/pubsub"
 )
 
 // ListenTask create a stream that will send data for every task to execute
 func (s *Server) ListenTask(request *ListenTaskRequest, stream Service_ListenTaskServer) (err error) {
-	subscription := pubsub.Subscribe(request.Service.TaskSubscriptionChannel())
+	service, err := services.Get(request.Token)
+	if err != nil {
+		return
+	}
+	subscription := pubsub.Subscribe(service.TaskSubscriptionChannel())
 	for data := range subscription {
 		execution := data.(*execution.Execution)
 		inputs, _ := json.Marshal(execution.Inputs)